Use native Android unit testing instead of Robolectric
Fixes #10

Caveats: I had to remove a test case which was testing a fragment to forward callbacks to a wrapped dispatcher. We can’t test anything that is part of the Android framework without Robolectric. @lungos and I thought that the test wasn’t overly valuable, so this seemed a reasonable price to pay considering the largely improved simplicity of the build.